pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "supabase (>=2.28.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "clerk-backend-api (>=5.0.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
